chg nRefs from long to int (further removing MS legacy)
git-svn-id: http://vtp.googlecode.com/svn/trunk@5571 2ba57974-ef2e-0410-970c-f3cb55f6a86a

+++ TerrainSDK/vtdata/vtString.h
@@ -37,10 +37,9 @@ typedef const unsigned char *pcuchar;
 
 struct vtStringData
 {
-	long nRefs;			// reference count
+	int nRefs;			// reference count
 	int nDataLength;	// length of data (including terminator)
 	int nAllocLength;	// length of allocation
-	// char data[nAllocLength]
 
 	char* data()		// char* to managed data
 		{ return (char*)(this+1); }
